Fix PHP notices when given an empty route
I don't know if this is the best or even correct fix, but in the web services project we need to get a match on an empty route, ie. "/", but the router currently throws a PHP notice on line 112 because $segment[0] is undefined at that point (because $segment is an empty string).

This PR stops the PHP notices but is there a better way?